DE157998;- Fix for clear data and crosstab refresh on column remove

+++ production/src/office/office-api-helper.js
@@ -223,12 +223,13 @@ export class OfficeApiHelper {
    *
    * @param {Office} context Excel Context
    * @param {Object} object Contains information obout the object
+   * @param {Boolean} isClear specift if object should be cleared or deleted
    * @memberof OfficeApiHelper
    */
-  deleteObjectTableBody = async (context, object) => {
-    const {isCrosstab, crosstabHeaderDimensions} = object;
+  deleteObjectTableBody = async (context, object, isClear) => {
+    const { isCrosstab, crosstabHeaderDimensions } = object;
     const tableObject = context.workbook.tables.getItem(object.bindId);
-    await this.deleteExcelTable(tableObject, context, isCrosstab, crosstabHeaderDimensions);
+    await this.deleteExcelTable(tableObject, context, isCrosstab, crosstabHeaderDimensions, isClear);
   }
 
   /**
@@ -377,9 +378,8 @@ export class OfficeApiHelper {
    * @return {Object}
    */
   getCrosstabRangeSafely = async (table, headerDimensions, context) => {
-    const {columnsY, rowsX} = headerDimensions;
-    const validColumnsY = await this.getValidOffset(table, columnsY, 'getRowsAbove', context);
-    const validRowsX = await this.getValidOffset(table, rowsX, 'getColumnsBefore', context);
+    const { columnsY, rowsX } = headerDimensions;
+    const { validColumnsY, validRowsX } = await this.getCrosstabHeadersSafely(table, columnsY, context, rowsX);
     const startingCell = table.getRange().getCell(0, 0).getOffsetRange(-validColumnsY, -validRowsX);
     return startingCell.getBoundingRect(table.getRange());
   }
@@ -417,31 +417,43 @@ export class OfficeApiHelper {
    */
   clearCrosstabRange = async (officeTable, crosstabHeaderDimensions, prevheaderDimensions, isCrosstab, context) => {
     try {
-      // Row headers
-      const leftRange = officeTable.getRange().getColumnsBefore(prevheaderDimensions.rowsX);
-      context.trackedObjects.add(leftRange);
+      let leftRange;
+      let topRange;
+      let titlesRange;
+
+      if (prevheaderDimensions.rowsX) {
+        // Row headers
+        leftRange = officeTable.getRange().getColumnsBefore(prevheaderDimensions.rowsX);
+        context.trackedObjects.add(leftRange);
+        // Title headers
+        titlesRange = officeTable.getRange().getCell(0, 0).getOffsetRange(0, -1).getResizedRange(-(prevheaderDimensions.columnsY), -(prevheaderDimensions.rowsX - 1));
+        context.trackedObjects.add(titlesRange);
+      }
+
       // Column headers
-      const topRange = officeTable.getRange().getRowsAbove(prevheaderDimensions.columnsY);
-      context.trackedObjects.add(topRange);
-      // Title headers
-      const titlesRange = officeTable
-        .getRange()
-        .getCell(0, 0)
-        .getOffsetRange(0, -1)
-        .getResizedRange(-(prevheaderDimensions.columnsY), -(prevheaderDimensions.rowsX - 1));
-      context.trackedObjects.add(titlesRange);
+      if (prevheaderDimensions.columnsY) {
+        topRange = officeTable.getRange().getRowsAbove(prevheaderDimensions.columnsY);
+        context.trackedObjects.add(topRange);
+      }
       // Check if ranges are valid before clearing
       await context.sync();
       if (isCrosstab && (crosstabHeaderDimensions === prevheaderDimensions)) {
-        leftRange.clear('contents');
-        topRange.clear('contents');
-        titlesRange.clear('contents');
+        if (prevheaderDimensions.columnsY) topRange.clear('contents');
+        if (prevheaderDimensions.rowsX) {
+          leftRange.clear('contents');
+          titlesRange.clear('contents');
+        }
       } else {
-        leftRange.clear();
-        topRange.clear();
-        titlesRange.clear();
+        if (prevheaderDimensions.columnsY) topRange.clear();
+        if (prevheaderDimensions.rowsX) {
+          leftRange.clear();
+          titlesRange.clear();
+        }
       }
-      context.trackedObjects.remove([leftRange, topRange, titlesRange]);
+      if (prevheaderDimensions.columnsY) context.trackedObjects.remove([topRange]);
+      if (prevheaderDimensions.rowsX) {
+        context.trackedObjects.remove([leftRange, titlesRange]);
+      }
     } catch (error) {
       officeTable.showHeaders = false;
       throw error;
@@ -457,8 +469,8 @@ export class OfficeApiHelper {
    * @memberof OfficeApiHelper
    * @return {Object}
    */
-  getTableStartCell = ({ startCell, instanceDefinition, prevOfficeTable, toCrosstabChange, fromCrosstabChange }) => {
-    const { mstrTable: { isCrosstab, prevCrosstabDimensions, crosstabHeaderDimensions } } = instanceDefinition;
+  getTableStartCell = ({ startCell, instanceDefinition, prevOfficeTable }) => {
+    const { mstrTable: { isCrosstab, prevCrosstabDimensions, crosstabHeaderDimensions, toCrosstabChange, fromCrosstabChange, } } = instanceDefinition;
     if (fromCrosstabChange) {
       return this.offsetCellBy(startCell, -prevCrosstabDimensions.columnsY, -prevCrosstabDimensions.rowsX);
     }
@@ -509,7 +521,6 @@ export class OfficeApiHelper {
     const directionVector = [1, 0];
     const headerRange = startingCell.getResizedRange(columns.length - 1, columns[0].length - 1);
     this.insertHeadersValues(headerRange, columns, 'columns');
-
     return this.createHeaders(columns, startingCell, directionVector);
   }
 
@@ -539,15 +550,31 @@ export class OfficeApiHelper {
   }
 
   /**
+  * Returns the number of rows and columns headers that are valid for crosstab
+  *
+  * @param {Office} table Excel Object containig information about Excel Table
+  * @param {Object} columnsY Contains information about crosstab header columnsY direction
+  * @param {Office} context Excel context
+  * @param {Object} rowsX Contains information about crosstab header rowsX direction
+  * @memberof OfficeApiHelper
+  */
+  async getCrosstabHeadersSafely(table, columnsY, context, rowsX) {
+    const validColumnsY = await this.getValidOffset(table, columnsY, 'getRowsAbove', context);
+    const validRowsX = await this.getValidOffset(table, rowsX, 'getColumnsBefore', context);
+    return { validColumnsY, validRowsX };
+  }
+
+  /**
   * Delete Excel table object from workbook. For crosstab reports will also clear the headers
   *
   * @param {Office} tableObject Address of the first cell in report (top left)
   * @param {Office} context Contains arrays of attributes names in crosstab report
   * @param {Boolean} isCrosstab Specify if object is a crosstab
   * @param {Object} crosstabHeaderDimensions Contains dimensions of crosstab report headers
+  * @param {Boolean} isClear specift if object should be cleared or deleted
   * @memberof OfficeApiHelper
   */
-  async deleteExcelTable(tableObject, context, isCrosstab = false, crosstabHeaderDimensions = {}) {
+  async deleteExcelTable(tableObject, context, isCrosstab = false, crosstabHeaderDimensions = {}, isClear = false) {
     context.runtime.enableEvents = false;
     await context.sync();
     const tableRange = tableObject.getDataBodyRange();
@@ -566,7 +593,7 @@ export class OfficeApiHelper {
       rowsHeaders.clear();
     }
     tableRange.clear();
-    tableObject.delete();
+    if (!isClear) tableObject.delete();
     context.runtime.enableEvents = true;
     await context.sync();
     context.trackedObjects.remove(tableRange);
